+++ tests/unit/test_sensors_validity.py
@@ -1,19 +1,16 @@
 import firebench.sensors as fbs
-import numpy as np
 import pytest
 
 
-def test_CS506_rms_outside_validity_range():
+def test_CS506_cl_outside_validity_range():
     with pytest.raises(ValueError, match="Fuel moisture value must be between 0 and 50 percent"):
-        a = fbs.CS506_rms(-1, None)
-        print(a)
+        fbs.CS506_cl(-1, 68)
     with pytest.raises(ValueError, match="Fuel moisture value must be between 0 and 50 percent"):
-        fbs.CS506_rms(50.1, None)
+        fbs.CS506_cl(50.1, 68)
+    with pytest.raises(ValueError, match="Confidence level cl must be between 0 and 100"):
+        fbs.CS506_cl(30, -1)
 
 
-def test_CS506_range90_outside_validity_range():
-    with pytest.raises(ValueError, match="Fuel moisture value must be between 0 and 50 percent"):
-        a = fbs.CS506_range90(-1, None)
-        print(a)
-    with pytest.raises(ValueError, match="Fuel moisture value must be between 0 and 50 percent"):
-        fbs.CS506_range90(50.1, None)
+def test_z_from_cl_outside_validity_range():
+    with pytest.raises(ValueError, match="Confidence level cl must be between 0 and 1"):
+        fbs.z_from_cl(-1)